requirements.txt
playwright
beautifulsoup4
lxml
google-generativeai
tzdata
